Testes e correção de inicialização da API
- Realizei o teste de login e autenticação (funcionando);
- Corrigi erros pequenos de inicialização da API;
- Iniciei a implementação da atualização do estoque com base na entrada e saída.

+++ API_CRUDBackend/src/main/java/school/sptech/CRUDBackend/dto/SaidaEstoque/SaidaEstoqueCadastroDto.java
@@ -16,6 +16,6 @@ public class SaidaEstoqueCadastroDto {
     private Integer idSaida;
     private LocalDate data;
     private LocalTime hora;
-    private Integer qtdSaida;
+    private Double qtdSaida;
     private String motivoSaida;
 }
